Move Tailwind installation to build script
