+++ Containerfile
@@ -26,10 +26,10 @@ RUN --mount=type=cache,target=/root/.cache/uv \
 
 # Install javascript dependencies with Bun
 # Uses a persistent cache mount for Bun
-COPY package.json bun.lock* ${PROJ_DIR}
+COPY package.json build.ts bun.lock* ${PROJ_DIR}
 RUN --mount=type=cache,target=/root/.bun/install/cache \
     bun install --frozen-lockfile && \
-    bun run install-tailwind
+    bun run build -- --only-tailwind
 
 # Copy application code to image
 COPY ./app ${APP_DIR}
